Change PSA_MAX_PERSISTENT_KEY_IDENTIFIER to 64 bits when using ITS
When compiling for SPM with ITS backend, the value of
PSA_MAX_PERSISTENT_KEY_IDENTIFIER must be a 64 bit value instead of the
default 32 bit. ITS UIDs are 64 bits and are expected to contain the
partition id (32 bits) in the lower 32 bits, and the psa_key_id_t
(32 bits) in the upper 32 bits of the ITS UID. In SPM, the NSPE
partition id is defined as 0xffffffff, and since the crypto
implementation reserves a range (0xffff0000-0xffffffff) of key ids for
internal/future use, crypto key ids must be smaller than 0xffff000000000000.